requirements.txt
flask
discord.py
requests
astral
python-dotenv
gunicorn
